fix: FashionDNN 모델 위치 수정
